requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-filter==23.3
django-cors-headers==4.3.0
django-storages[s3]==1.14.2
django-ratelimit==4.1.0
celery[redis]==5.3.6
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
uvicorn[standard]==0.24.0.post1
sentry-sdk==1.35.0
django-prometheus==2.3.1
python-dotenv==1.0.0
boto3==1.28.85
requests==2.31.0
whitenoise==6.6.0
Pillow==10.1.0
